Load CSS inline instead of externally
to quickly fix issue [1] where CSS intermittently fails to load in CMS/LMS.

Background:
===========
For some unknown reason, the external stylesheet fails to load in _one_
of the LMS [2] or CMS [3]; though the failure alternates between the
two. So, it will work in CMS but fail in LMS, then the success/failure
switches after several minutes, alternating back and forth. When this
happens, the same computed hash value is used for each service's
resource.

Until additional resources can be devoted to investigating the root
cause, let's simply load the stylesheet inline;
the file size is 3.1K.

References:
===========
- [1] TNL-7084
- [2] HTTP 403 via CDN
- [3] HTTP 404 via nginx